+++ src/themes/index.js
@@ -1,4 +1,4 @@
-import styled, { keyframes } from 'styled-components';
+import styled from 'styled-components';
 import invertColors from '../utils/invertColors';
 
 import material from './material';
@@ -55,24 +55,3 @@ export const getStyles = (styles, component, multiple) => styled(component || 'd
    ${props => styles[props.theme.type](props)}
  `;
 */
-
-
-export const spin = keyframes`
-  to { transform: rotate(1turn); }
-`;
-export const spinner = (theme) => `
-  animation: ${spin} 400ms infinite linear;
-  width: ${theme.spinnerSize}px;
-  height: ${theme.spinnerSize}px;
-  box-sizing: border-box;
-  border-radius: 50%;
-  border: ${Math.floor(theme.spinnerSize / 8)}px solid ${theme.base05};
-  border-right-color: ${theme.base01};
-  display: inline-block;
-  position: relative;
-`;
-
-export const fadeIn = keyframes`
-  from { opacity: 0; }
-  to { opacity: 1; }
-`;
