+++ src/utils/effects.js
@@ -1,0 +1,52 @@
+import { keyframes } from 'styled-components';
+
+export const spin = keyframes`
+  to { transform: rotate(1turn); }
+`;
+export const spinner = (theme) => `
+  animation: ${spin} 400ms infinite linear;
+  width: ${theme.spinnerSize}px;
+  height: ${theme.spinnerSize}px;
+  box-sizing: border-box;
+  border-radius: 50%;
+  border: ${Math.floor(theme.spinnerSize / 8)}px solid ${theme.base05};
+  border-right-color: ${theme.base01};
+  display: inline-block;
+  position: relative;
+`;
+
+export const fadeIn = keyframes`
+  from { opacity: 0; }
+  to { opacity: 1; }
+`;
+
+// Based on https://github.com/mladenplavsic/css-ripple-effect
+export const ripple = (theme) => `
+  & {
+    position: relative;
+    overflow: hidden;
+  }
+
+  &:after {
+    content: "";
+    display: block;
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    pointer-events: none;
+    background-image: radial-gradient(circle, ${theme.base00} 10%, transparent 10.01%);
+    background-repeat: no-repeat;
+    background-position: 50%;
+    transform: scale(10 ,10);
+    opacity: 0;
+    transition: transform .5s, opacity 1s;
+  }
+
+  &:active:after {
+    transform: scale(0, 0);
+    opacity: .2;
+    transition: 0s;
+  }
+`;
